+++ AmplifyPlugins/DataStore/Tests/AWSDataStorePluginTests/Sync/Support/MutationEventExtensionsTests.swift
@@ -37,7 +37,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
                                                            version: nil)
         let responseMutationSync = createMutationSync(model: post, version: 1)
 
-        setUpPendingMutationQueue(modelId, [requestMutationEvent, pendingMutationEvent], pendingMutationEvent)
+        setUpPendingMutationQueue(post, [requestMutationEvent, pendingMutationEvent], pendingMutationEvent)
 
         let reconciledEvent = MutationEvent.reconcile(pendingMutationEvent: pendingMutationEvent,
                                                       with: requestMutationEvent,
@@ -62,7 +62,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
         wait(for: [updatingVersionExpectation], timeout: 1)
 
         // query for head of mutation event table for given model id and check if it has the updated version
-        MutationEvent.pendingMutationEvents(for: post.id,
+        MutationEvent.pendingMutationEvents(forModel: post,
                                             storageAdapter: storageAdapter) { result in
             switch result {
             case .failure(let error):
@@ -106,7 +106,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
                                                                  version: nil)
         let responseMutationSync = createMutationSync(model: post, version: 1)
 
-        setUpPendingMutationQueue(modelId,
+        setUpPendingMutationQueue(post,
                                   [requestMutationEvent, pendingUpdateMutationEvent, pendingDeleteMutationEvent],
                                   pendingUpdateMutationEvent)
 
@@ -133,7 +133,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
         wait(for: [updatingVersionExpectation], timeout: 1)
 
         // query for head of mutation event table for given model id and check if it has the updated version
-        MutationEvent.pendingMutationEvents(for: post.id,
+        MutationEvent.pendingMutationEvents(forModel: post,
                                             storageAdapter: storageAdapter) { result in
             switch result {
             case .failure(let error):
@@ -173,7 +173,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
                                                            version: 2)
         let responseMutationSync = createMutationSync(model: post1, version: 1)
 
-        setUpPendingMutationQueue(modelId, [requestMutationEvent, pendingMutationEvent], pendingMutationEvent)
+        setUpPendingMutationQueue(post1, [requestMutationEvent, pendingMutationEvent], pendingMutationEvent)
 
         let reconciledEvent = MutationEvent.reconcile(pendingMutationEvent: pendingMutationEvent,
                                                       with: requestMutationEvent,
@@ -197,7 +197,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
         wait(for: [updatingVersionExpectation], timeout: 1)
 
         // query for head of mutation event table for given model id and check if it has the correct version
-        MutationEvent.pendingMutationEvents(for: post1.id,
+        MutationEvent.pendingMutationEvents(forModel: post1,
                                             storageAdapter: storageAdapter) { result in
             switch result {
             case .failure(let error):
@@ -237,7 +237,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
                                                            version: 1)
         let responseMutationSync = createMutationSync(model: post3, version: 2)
 
-        setUpPendingMutationQueue(modelId, [requestMutationEvent, pendingMutationEvent], pendingMutationEvent)
+        setUpPendingMutationQueue(post1, [requestMutationEvent, pendingMutationEvent], pendingMutationEvent)
 
         let reconciledEvent = MutationEvent.reconcile(pendingMutationEvent: pendingMutationEvent,
                                                       with: requestMutationEvent,
@@ -261,7 +261,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
         wait(for: [updatingVersionExpectation], timeout: 1)
 
         // query for head of mutation event table for given model id and check if it has the correct version
-        MutationEvent.pendingMutationEvents(for: post1.id,
+        MutationEvent.pendingMutationEvents(forModel: post1,
                                             storageAdapter: storageAdapter) { result in
             switch result {
             case .failure(let error):
@@ -300,7 +300,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
                                                            version: 1)
         let responseMutationSync = createMutationSync(model: post1, version: 2)
 
-        setUpPendingMutationQueue(modelId, [requestMutationEvent, pendingMutationEvent], pendingMutationEvent)
+        setUpPendingMutationQueue(post1, [requestMutationEvent, pendingMutationEvent], pendingMutationEvent)
 
         let reconciledEvent = MutationEvent.reconcile(pendingMutationEvent: pendingMutationEvent,
                                                       with: requestMutationEvent,
@@ -324,7 +324,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
         wait(for: [updatingVersionExpectation], timeout: 1)
 
         // query for head of mutation event table for given model id and check if it has the correct version
-        MutationEvent.pendingMutationEvents(for: post1.id,
+        MutationEvent.pendingMutationEvents(forModel: post1,
                                             storageAdapter: storageAdapter) { result in
             switch result {
             case .failure(let error):
@@ -366,7 +366,7 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
         return MutationSync(model: AnyModel(model), syncMetadata: metadata)
     }
 
-    private func setUpPendingMutationQueue(_ modelId: String,
+    private func setUpPendingMutationQueue(_ model: Model,
                                            _ mutationEvents: [MutationEvent],
                                            _ expectedHeadOfQueue: MutationEvent) {
         for mutationEvent in mutationEvents {
@@ -383,8 +383,10 @@ class MutationEventExtensionsTest: BaseDataStoreTests {
 
         // verify the head of queue is expected
         let headOfQueueExpectation = expectation(description: "head of mutation event queue is as expected")
-        MutationEvent.pendingMutationEvents(for: modelId,
-                                            storageAdapter: storageAdapter) { result in
+        MutationEvent.pendingMutationEvents(
+            forModel: model,
+            storageAdapter: storageAdapter
+        ) { result in
             switch result {
             case .failure(let error):
                 XCTFail("Error : \(error)")